[TASK-199] 이미지 수정
