+++ src/assets/images/index.ts
@@ -109,3 +109,17 @@ export { ReactComponent as Error404 } from './error_404.svg';
 
 // ? 여행방 썸네일
 export { ReactComponent as UserCount } from './user_count.svg';
+
+export { ReactComponent as MiniIconC1 } from './mini_icon_c1.svg';
+export { ReactComponent as MiniIconC2 } from './mini_icon_c2.svg';
+export { ReactComponent as MiniIconC3 } from './mini_icon_c3.svg';
+export { ReactComponent as MiniIconC4 } from './mini_icon_c4.svg';
+export { ReactComponent as MiniIconC5 } from './mini_icon_c5.svg';
+export { ReactComponent as MiniIconC6 } from './mini_icon_c6.svg';
+
+export { ReactComponent as MiniIconP1 } from './mini_icon_p1.svg';
+export { ReactComponent as MiniIconP2 } from './mini_icon_p2.svg';
+export { ReactComponent as MiniIconP3 } from './mini_icon_p3.svg';
+export { ReactComponent as MiniIconP4 } from './mini_icon_p4.svg';
+export { ReactComponent as MiniIconP5 } from './mini_icon_p5.svg';
+export { ReactComponent as MiniIconP6 } from './mini_icon_p6.svg';
