+++ src/components/MainHeader/MainHeader.module.scss
@@ -7,4 +7,18 @@
 
   padding: 10px auto;
   border-bottom: 1px solid #e8e8e8;
+  position: relative;
+
+  .sideLogo {
+    position: absolute;
+    @include flex(center, center);
+    width: 40px;
+    height: 40px;
+    border-radius: 999px;
+    border: 1px solid $gray30;
+    background-color: $gray10;
+
+    right: 21px;
+    top: 14px;
+  }
 }
