+++ src/pages/Home/Home.tsx
@@ -13,10 +13,6 @@ import TRAVEL_ROOM_THUMBNAIL3 from '@images/travel_room_thumbnail3.png';
 import TRAVEL_ROOM_THUMBNAIL4 from '@images/travel_room_thumbnail4.png';
 import TRAVEL_ROOM_THUMBNAIL5 from '@images/travel_room_thumbnail5.png';
 import TRAVEL_ROOM_THUMBNAIL6 from '@images/travel_room_thumbnail6.png';
-import { useEffect } from 'react';
-import { tentenInstance } from '@src/constants/axios';
-import userDataAtom from '@src/recoil/userData/atom';
-import { useSetRecoilState } from 'recoil';
 
 const THUMBNAIL_IMAGES = [
   TRAVEL_ROOM_THUMBNAIL1,
@@ -28,7 +24,6 @@ const THUMBNAIL_IMAGES = [
 ];
 
 const Home = () => {
-  const setUserData = useSetRecoilState(userDataAtom);
   const { travelRoomData } = useGetTravelRoom();
 
   const navigate = useNavigate();
@@ -36,20 +31,6 @@ const Home = () => {
   const handleCreateTravel = () => {
     navigate(ROUTES.createTravel);
   };
-
-  useEffect(() => {
-    const fetchUserData = async () => {
-      try {
-        const response = await tentenInstance.get('/users/current');
-
-        setUserData(response.data);
-      } catch (error) {
-        console.error(error);
-      }
-    };
-
-    fetchUserData();
-  }, []);
 
   // ? 여행 방이 없을 때
   if (!travelRoomData || travelRoomData.length < 1) {
